Time: 0 ms (100.00%) | Memory: 8.1 MB (87.46%) - LeetSync

+++ 172-factorial-trailing-zeroes/factorial-trailing-zeroes.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int trailingZeroes(int n) {
         int result = 0;
-        for(long i=5;n/i>0;i*=5){
+        for(int i=5;n/i>0;i*=5){
             result+=(n/i);
         }
         return result;
